# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SME Finance Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /profiles/{profileId}/transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List transactions for a profile
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: List of transactions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Create a manual transaction
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Transaction created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /profiles/{profileId}/transactions/{transactionId}:
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get one Financial Memory transaction for a profile
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction or profile not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Transactions]
      operationId: updateTransaction
      summary: Edit a manually added transaction
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Transaction updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    delete:
      tags: [Transactions]
      operationId: deleteTransaction
      summary: Delete a manually added transaction
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Transaction deleted.
...
components:
  parameters:
...
    ProfileIdParam:
      name: profileId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...
    Transaction:
      type: object
      required: [id, profileId, date, description, amount, recordType, category, taxTreatment, source, createdAt, updatedAt]
      properties:
        id:
          type: string
        profileId:
          type: string
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        recordType:
          type: string
          enum: [income, expense, unknown]
        category:
          type: string
        note:
          type: [string, 'null']
        taxTreatment:
          type: string
        source:
          type: string
        evidenceId:
          type: [string, 'null']
        accountingClassification:
          type: [string, 'null']
          enum: [income, expense, transfer, owner_funds, drawings, loan, tax_payment, unknown, null]
        financialAccountId:
          type: [string, 'null']
        bankImportBatchId:
          type: [string, 'null']
        bankImportRowId:
          type: [string, 'null']
        bankMovementIdentity:
          type: [string, 'null']
        ledgerStatus:
          type: string
          enum: [active, voided]
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionInput:
      type: object
      required: [date, recordType, description, amount, category]
      properties:
        date:
          type: string
        recordType:
          type: string
          enum: [income, expense]
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        note:
          type: [string, 'null']
    TransactionUpdate:
      type: object
      properties:
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        taxTreatment:
          type: string
        accountingClassification:
          type: string
          enum: [income, expense, transfer, owner_funds, drawings, loan, tax_payment, unknown]
...