# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SME Finance Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUpdateAccountingClassification } from './transactionUpdateAccountingClassification';

export interface TransactionUpdate {
  date?: string;
  description?: string;
  amount?: number;
  category?: string;
  taxTreatment?: string;
  accountingClassification?: TransactionUpdateAccountingClassification;
}